Add solutions for exercises 4.15 to 4.23

+++ part4/bloglist-backend/controllers/blogs.js
@@ -1,6 +1,6 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
-const User = require("../models/user");
+const middleware = require("../utils/middleware");
 
 blogsRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({}).populate("user");
@@ -16,30 +16,48 @@ blogsRouter.get("/:id", async (request, response) => {
   }
 });
 
-blogsRouter.post("/", async (request, response) => {
+blogsRouter.post("/", middleware.userExtractor, async (request, response) => {
   const body = request.body;
+  const user = request.user;
 
-  // const user = await User.findById(body.userId);
+  if (!user) {
+    return response.status(401).json({ error: "invalid user" });
+  }
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes,
-    // user: user.id,
+    user: user._id,
   });
   const savedBlog = await blog.save();
 
-  // user.blogs = user.blogs.concat(savedBlog._id);
-  // await user.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
 
   response.status(201).json(savedBlog);
 });
 
-blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
-});
+blogsRouter.delete(
+  "/:id",
+  middleware.userExtractor,
+  async (request, response) => {
+    const blog = await Blog.findById(request.params.id);
+    if (!blog) {
+      return response.status(204).end();
+    }
+
+    if (!request.user || request.user.id.toString() !== blog.user.toString()) {
+      return response
+        .status(401)
+        .json({ error: "user does not have permission to delete this blog" });
+    }
+
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end();
+  }
+);
 
 blogsRouter.put("/:id", async (request, response) => {
   const { title, author, url, likes } = request.body;
